fix: match vendor float32, reset patches on exec, SKIP_MEGA_BATCH_EVAL, dtype print
- Harness: no torch_dtype arg (float32, matches vendor from_pretrained)
- Eval scripts: no torch_dtype arg (same)
- YAML: git checkout vendor/baselines before apply_all.py (forces re-patch)
- All baselines: SKIP_MEGA_BATCH_EVAL=1, ORDERING required
- source_patches: prints Model dtype after loading
- .half() → .to(param.dtype) in experiment scripts
- Clean dtype test suite (test_vendor_dtype_match.py)

+++ src/experiments/installation_strength.py
@@ -156,7 +156,7 @@ def eval_at_1k_checkpoint(seed: int, ckpt_base: Path, output_dir: Path, model_na
         param_dict = dict(model.named_parameters())
         for name, tensor in weights.items():
             if name in param_dict:
-                param_dict[name].data.copy_(tensor.cuda().half())
+                param_dict[name].data.copy_(tensor.cuda().to(param_dict[name].dtype))
         del weights
         torch.cuda.empty_cache()
